requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
exceptiongroup==1.0.4
fastapi==0.87.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
pydantic==1.10.2
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.21.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
undetected-chromedriver==3.1.6
urllib3==1.26.12
uvicorn==0.19.0
websockets==10.4
wsproto==1.2.0
